requirements.txt
fastapi
uvicorn
sqlmodel
aiosqlite
google-generativeai
python-multipart
python-dotenv
pandas
numpy
pillow
httpx
jinja2
pydantic-settings
